pyproject.toml
[tool.poetry]
name = "lms-api"
version = "0.1.0"
description = ""
authors = ["Goldyban Nastya <nasty-goldyba@ya.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
djangorestframework = "^3.15.2"
psycopg2 = "^2.9.10"
python-dotenv = "^1.0.1"
pillow = "^11.1.0"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.4.0"
coverage = "^7.6.10"
drf-spectacular = "^0.28.0"
stripe = "^11.5.0"


[tool.poetry.group.lint.dependencies]
black = "^24.10.0"
isort = "^5.13.2"
flake8 = "^7.1.1"

[tool.black]
line-length = 119
exclude = '.git'

[tool.isort]
line_length = 119

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
